Test: adding test(VenteTest and HistoricalTest)
