+++ src/main/java/school/hei/patrimoine/modele/vente/ValeurMarche.java
@@ -2,6 +2,7 @@ package school.hei.patrimoine.modele.vente;
 
 import java.io.Serializable;
 import java.time.LocalDate;
+
 import school.hei.patrimoine.modele.Argent;
 import school.hei.patrimoine.modele.possession.Possession;
