+++ src/test/java/school/hei/patrimoine/modele/vente/ValeurMarcheHistoriqueTest.java
@@ -1,8 +1,5 @@
 package school.hei.patrimoine.modele.vente;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import java.time.LocalDate;
 import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,6 +8,8 @@ import school.hei.patrimoine.modele.Argent;
 import school.hei.patrimoine.modele.Devise;
 import school.hei.patrimoine.modele.possession.Compte;
 import school.hei.patrimoine.modele.possession.Materiel;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ValeurMarcheHistoriqueTest {
   private Materiel materiel;
@@ -39,6 +38,7 @@ public class ValeurMarcheHistoriqueTest {
 
     assertEquals(1, historique.size());
     ValeurMarche premiereValeur = historique.iterator().next();
+    assertEquals(materiel, premiereValeur.possession());
     assertEquals(aujourdhui, premiereValeur.t());
     assertEquals(valeurInitiale, premiereValeur.valeur());
   }
@@ -47,22 +47,23 @@ public class ValeurMarcheHistoriqueTest {
   void historique_valeur_marche_apres_ajout_nouvelle_valeur() {
     LocalDate demain = aujourdhui.plusDays(1);
     Argent nouvelleValeur = new Argent(1050, Devise.EUR);
-    materiel.ajouterValeurMarche(new ValeurMarche(demain, nouvelleValeur));
+    materiel.ajouterValeurMarche(new ValeurMarche(materiel, demain, nouvelleValeur));
 
     Set<ValeurMarche> historique = materiel.historiqueValeurMarche();
 
     assertEquals(2, historique.size());
-    assertTrue(historique.contains(new ValeurMarche(aujourdhui, valeurInitiale)));
-    assertTrue(historique.contains(new ValeurMarche(demain, nouvelleValeur)));
+    assertTrue(historique.contains(new ValeurMarche(materiel, aujourdhui, valeurInitiale)));
+    assertTrue(historique.contains(new ValeurMarche(materiel, demain, nouvelleValeur)));
   }
 
   @Test
   void historique_valeur_marche_pour_type_non_eligible() {
     Compte compte = new Compte("Compte Courant", aujourdhui, valeurInitiale);
 
-    Set<ValeurMarche> historique = compte.historiqueValeurMarche();
+    assertThrows(UnsupportedOperationException.class, () -> {
+      compte.ajouterValeurMarche(new ValeurMarche(compte, aujourdhui, valeurInitiale));
+    });
 
-    assertEquals(1, historique.size());
-    assertEquals(valeurInitiale, historique.iterator().next().valeur());
+    assertEquals(0, compte.historiqueValeurMarche().size());
   }
 }
